batch5: make the borrowed patches pass the checks they never ran
#325 and #391 share no merge base with main, so their CI never produced a
verdict on this tree. Both needed work their PRs did not contain:

  * `find_args_start`'s new gemma4 arm was a nested `if let`, which this
    workspace denies (collapsible_if). Rewritten as a let-chain, matching
    the Qwen3-Coder arm directly below it.
  * `streaming_emit.rs` was not rustfmt-clean.

Also splits `app_input_tests.rs`, which #489's routing test pushed to 521
lines and over the CI cap -- that check is currently RED on #489 itself. The
Help routing tests move to `app_input_help_tests.rs`; the three fixtures they
share become `pub(super)` rather than being copied, the convention
`render_tests`/`start_tests` already use.
